Added code for JWT_extended

+++ resources/item.py
@@ -1,5 +1,5 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import jwt_required
 from models.item import ItemModel
 
 class Item(Resource):
@@ -15,7 +15,7 @@ class Item(Resource):
 		help="Every item must have a store id."
 	)
 
-	@jwt_required()	# add this decorator to any method you want authorised via the session token
+	@jwt_required	# add this decorator to any method you want authorised via the session token
 	def get(self, name):
 		try:
 			item = ItemModel.find_by_name(name)
@@ -41,11 +41,15 @@ class Item(Resource):
 
 		return item.json(), 201		# 201 is CREATED http status code
 					
+	@jwt_required
 	def delete(self, name):
+		claims = get_jwt_claims()
+		if not claims['is_admin']:
+			return {'message': 'Admin priviledges required.'}, 401
+
 		item = ItemModel.find_by_name(name)
 		if item:
 			item.delete_from_db()
-		
 		return {'message': 'Item deleted'}
 	
 	def put(self, name):
